gws: reorganizes testing related modules
