+++ src/sync/gws/mod.rs
@@ -1,4 +1,6 @@
 mod api;
+#[cfg(test)]
+mod tests;
 
 use crate::sync::gws::api::{GoogleWorkspaceApiClient, Group, User, UserName};
 use std::collections::BTreeSet;
@@ -138,327 +140,5 @@ impl SyncGoogleWorkspace {
             .collect();
 
         Ok(diffs)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    pub mod rust_team_data_fakes {
-        use rust_team_data::v1::{GoogleWorkspace, Team, TeamKind, TeamMember};
-
-        pub fn normal_member(name: &str) -> TeamMember {
-            TeamMember {
-                name: name.into(),
-                github: name.into(),
-                github_id: 1234567,
-                is_lead: false,
-                roles: vec![],
-                google_workspace: None,
-            }
-        }
-
-        pub fn privileged_member(name: &str, surname: &str) -> TeamMember {
-            TeamMember {
-                google_workspace: Some(GoogleWorkspace {
-                    first_name: name.into(),
-                    last_name: surname.into(),
-                    account_handle: format!("{name}.{surname}"),
-                }),
-                ..normal_member(name)
-            }
-        }
-
-        pub fn normal_team(name: &str, members: Vec<TeamMember>) -> Team {
-            Team {
-                kind: TeamKind::Team,
-                name: name.to_string(),
-                github: None,
-                website_data: None,
-                subteam_of: None,
-                top_level: Some(true),
-                alumni: vec![],
-                roles: vec![],
-                google_workspace_saml_group: None,
-                members,
-            }
-        }
-
-        pub fn privileged_team(name: &str, members: Vec<TeamMember>) -> Team {
-            Team {
-                google_workspace_saml_group: Some(true),
-                ..normal_team(name, members)
-            }
-        }
-    }
-
-    use crate::sync::gws::api::{GoogleWorkspaceApiClient, Group, User, UserName};
-    use crate::sync::gws::tests::rust_team_data_fakes::{privileged_member, privileged_team};
-    use crate::sync::gws::{
-        GoogleGroupDiff, GoogleUserDiff, GoogleWorkspaceDiff, RUST_LANG_GWS_DOMAIN,
-        SyncGoogleWorkspace,
-    };
-    use async_trait::async_trait;
-    use rust_team_data::v1::Team;
-
-    struct FakeGoogleWorkspace {
-        users: Vec<User>,
-        groups: Vec<Group>,
-    }
-
-    #[async_trait]
-    impl GoogleWorkspaceApiClient for FakeGoogleWorkspace {
-        async fn get_users(&self) -> anyhow::Result<Vec<User>> {
-            Ok(self.users.clone())
-        }
-
-        async fn get_groups(&self) -> anyhow::Result<Vec<Group>> {
-            Ok(self.groups.clone())
-        }
-    }
-
-    fn google_user(name: &str, surname: &str) -> User {
-        User {
-            name: UserName {
-                given_name: name.into(),
-                family_name: surname.into(),
-            },
-            primary_email: format!("{name}.{surname}@{RUST_LANG_GWS_DOMAIN}"),
-        }
-    }
-
-    fn google_group(name: &str) -> Group {
-        Group {
-            name: name.to_string(),
-            email: format!("{name}@{RUST_LANG_GWS_DOMAIN}"),
-        }
-    }
-
-    async fn run_sync(
-        gws_api_client: Box<dyn GoogleWorkspaceApiClient>,
-        teams: Vec<Team>,
-    ) -> GoogleWorkspaceDiff {
-        let sync = SyncGoogleWorkspace::new(teams, gws_api_client)
-            .await
-            .expect("cannot create sync");
-
-        let google_users_diff = sync.diff_users().expect("cannot diff accounts");
-        let google_groups_diff = sync.diff_groups().expect("cannot diff groups");
-        GoogleWorkspaceDiff {
-            google_users: google_users_diff,
-            google_groups: google_groups_diff,
-        }
-    }
-
-    fn fake_gws_client(users: Vec<User>, groups: Vec<Group>) -> Box<dyn GoogleWorkspaceApiClient> {
-        let fake_gws = FakeGoogleWorkspace { users, groups };
-        Box::new(fake_gws)
-    }
-
-    #[tokio::test]
-    async fn diff_spots_nothing() {
-        let google_users = vec![
-            google_user("ubiratan", "soares"),
-            google_user("marco", "ieni"),
-        ];
-
-        let google_groups = vec![
-            google_group("infra-admins-saml"),
-            google_group("security-response"), // groups not related to SAML are not diffed
-        ];
-
-        let teams = vec![privileged_team(
-            "infra-admins",
-            vec![
-                privileged_member("ubiratan", "soares"),
-                privileged_member("marco", "ieni"),
-            ],
-        )];
-
-        let gws_api_client = fake_gws_client(google_users, google_groups);
-
-        let diff = run_sync(gws_api_client, teams).await;
-        assert!(diff.google_users.is_empty());
-        assert!(diff.google_groups.is_empty());
-    }
-
-    #[tokio::test]
-    async fn diff_spots_user_creation() {
-        let google_users = vec![
-            google_user("ubiratan", "soares"),
-            google_user("marco", "ieni"),
-        ];
-
-        let google_groups = vec![google_group("infra-admins-saml")];
-
-        let teams = vec![privileged_team(
-            "infra-admins",
-            vec![
-                privileged_member("ubiratan", "soares"),
-                privileged_member("marco", "ieni"),
-                privileged_member("emily", "albini"),
-            ],
-        )];
-
-        let gws_api_client = fake_gws_client(google_users, google_groups);
-
-        let diff = run_sync(gws_api_client, teams).await;
-        let expected = vec![GoogleUserDiff::Create(google_user("emily", "albini"))];
-
-        assert_eq!(diff.google_users, expected);
-        assert!(diff.google_groups.is_empty());
-    }
-
-    #[tokio::test]
-    async fn diff_spots_user_deletion() {
-        let google_users = vec![
-            google_user("ubiratan", "soares"),
-            google_user("marco", "ieni"),
-            google_user("emily", "albini"),
-        ];
-
-        let google_groups = vec![google_group("infra-admins-saml")];
-
-        let teams = vec![privileged_team(
-            "infra-admins",
-            vec![privileged_member("emily", "albini")],
-        )];
-
-        let gws_api_client = fake_gws_client(google_users, google_groups);
-
-        let diff = run_sync(gws_api_client, teams).await;
-        let expected = vec![
-            GoogleUserDiff::Delete(google_user("ubiratan", "soares")),
-            GoogleUserDiff::Delete(google_user("marco", "ieni")),
-        ];
-
-        assert_eq!(diff.google_users, expected);
-        assert!(diff.google_groups.is_empty());
-    }
-
-    #[tokio::test]
-    async fn diff_spots_group_creation() {
-        let google_users = vec![
-            google_user("ubiratan", "soares"),
-            google_user("marco", "ieni"),
-        ];
-
-        let google_groups = vec![];
-
-        let teams = vec![privileged_team(
-            "infra-admins",
-            vec![
-                privileged_member("ubiratan", "soares"),
-                privileged_member("marco", "ieni"),
-            ],
-        )];
-
-        let gws_api_client = fake_gws_client(google_users, google_groups);
-
-        let diff = run_sync(gws_api_client, teams).await;
-        let expected = vec![GoogleGroupDiff::Create(google_group("infra-admins-saml"))];
-
-        assert!(diff.google_users.is_empty());
-        assert_eq!(expected, diff.google_groups);
-    }
-
-    #[tokio::test]
-    async fn diff_spots_group_deletion() {
-        let google_users = vec![
-            google_user("ubiratan", "soares"),
-            google_user("marco", "ieni"),
-        ];
-
-        let google_groups = vec![
-            google_group("infra-admins-saml"),
-            google_group("external-auditors-saml"),
-        ];
-
-        let teams = vec![privileged_team(
-            "infra-admins",
-            vec![
-                privileged_member("ubiratan", "soares"),
-                privileged_member("marco", "ieni"),
-            ],
-        )];
-
-        let gws_api_client = fake_gws_client(google_users, google_groups);
-
-        let diff = run_sync(gws_api_client, teams).await;
-        let expected = vec![GoogleGroupDiff::Delete(google_group(
-            "external-auditors-saml",
-        ))];
-
-        assert!(diff.google_users.is_empty());
-        assert_eq!(expected, diff.google_groups);
-    }
-
-    #[tokio::test]
-    async fn diff_spots_multiple_creations() {
-        let google_users = vec![
-            google_user("ubiratan", "soares"),
-            google_user("marco", "ieni"),
-        ];
-
-        let google_groups = vec![google_group("infra-admins-saml")];
-
-        let teams = vec![
-            privileged_team(
-                "infra-admins",
-                vec![
-                    privileged_member("ubiratan", "soares"),
-                    privileged_member("marco", "ieni"),
-                ],
-            ),
-            privileged_team(
-                "crates-io-admins",
-                vec![privileged_member("adam", "harvey")],
-            ),
-        ];
-
-        let gws_api_client = fake_gws_client(google_users, google_groups);
-
-        let diff = run_sync(gws_api_client, teams).await;
-        let expected = GoogleWorkspaceDiff {
-            google_users: vec![GoogleUserDiff::Create(google_user("adam", "harvey"))],
-            google_groups: vec![GoogleGroupDiff::Create(google_group(
-                "crates-io-admins-saml",
-            ))],
-        };
-
-        assert_eq!(expected, diff);
-    }
-
-    #[tokio::test]
-    async fn diff_spots_multiple_deletions() {
-        let google_users = vec![
-            google_user("ubiratan", "soares"),
-            google_user("marco", "ieni"),
-            google_user("adam", "harvey"),
-        ];
-
-        let google_groups = vec![
-            google_group("infra-admins-saml"),
-            google_group("crates-io-admins-saml"),
-        ];
-
-        let teams = vec![privileged_team(
-            "infra-admins",
-            vec![
-                privileged_member("ubiratan", "soares"),
-                privileged_member("marco", "ieni"),
-            ],
-        )];
-
-        let gws_api_client = fake_gws_client(google_users, google_groups);
-
-        let diff = run_sync(gws_api_client, teams).await;
-        let expected = GoogleWorkspaceDiff {
-            google_users: vec![GoogleUserDiff::Delete(google_user("adam", "harvey"))],
-            google_groups: vec![GoogleGroupDiff::Delete(google_group(
-                "crates-io-admins-saml",
-            ))],
-        };
-
-        assert_eq!(expected, diff);
     }
 }
